TINY-3909: separate `removeTrailingBr` function to a separate file and call in both serializer and ParserFilters.ts
